Fixed up small bug in Status, and added tests.
There was a small bug in Status calculation, where it would not
set the applied_at, isApplied to default values after finding
the last applied migrations. So, migrations that were not in the
database table, would end up with the status of the last migration
in the database table, instead of pending.

Added test to make sure events returned from status match up with
expected events.

Added the ability to compare events to each other, mainly used in tests.